requirements.txt
django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
drf-spectacular==0.26.1
django-countries==7.5.1
django-countries[pyuca]